# blog-backend/apis/swagger/proto/user.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User API",
    "description": "User management API",
    "version": "v1.0"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/users/login": {
      "post": {
        "summary": "User login",
        "description": "User login",
        "operationId": "UserService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbLoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbLoginRequest"
            }
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/v1/users/logout": {
      "post": {
        "summary": "User logout",
        "description": "User logout",
        "operationId": "UserService_Logout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbLogoutResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbLogoutRequest"
            }
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/v1/users/register": {
      "post": {
        "summary": "User registration",
        "description": "User registration",
        "operationId": "UserService_Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbRegisterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbRegisterRequest"
            }
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "definitions": {
...
    "pbLoginRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "title": "用户名"
        },
        "password": {
          "type": "string",
          "title": "密码"
        }
      },
      "title": "用户登录请求"
    },
    "pbLoginResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/pbUser",
          "title": "用户信息"
        },
        "token": {
          "type": "string",
          "title": "token"
        }
      },
      "title": "用户登录响应"
    },
    "pbLogoutRequest": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string",
          "format": "int64",
          "title": "用户ID"
        }
      },
      "title": "用户登出请求"
    },
    "pbLogoutResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "用户登出响应"
    },
    "pbRegisterRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "title": "用户名"
        },
        "password": {
          "type": "string",
          "title": "密码"
        },
        "role": {
          "$ref": "#/definitions/pbRole",
          "title": "用户角色"
        }
      },
      "title": "用户注册请求（或者·叫用户新增请求，只有管理员才有权限）"
    },
    "pbRegisterResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/pbUser",
          "title": "用户信息"
        }
      },
      "title": "用户注册响应"
    },
...
    "pbRole": {
      "type": "string",
      "enum": [
        "ADMIN",
        "USER"
      ],
      "default": "ADMIN",
      "description": "- ADMIN: 管理员\n - USER: 普通用户",
      "title": "用户角色"
    },
    "pbStatus": {
      "type": "string",
      "enum": [
        "NORMAL",
        "DISABLED"
      ],
      "default": "NORMAL",
      "description": "- NORMAL: 正常\n - DISABLED: 禁用",
      "title": "用户状态"
    },
...
    "pbUser": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string",
          "format": "int64",
          "title": "用户ID"
        },
        "username": {
          "type": "string",
          "title": "用户名"
        },
        "email": {
          "type": "string",
          "title": "邮箱"
        },
        "phone": {
          "type": "string",
          "title": "手机号"
        },
        "role": {
          "$ref": "#/definitions/pbRole",
          "title": "用户角色"
        },
        "status": {
          "$ref": "#/definitions/pbStatus",
          "title": "用户状态"
        },
        "avatar": {
          "type": "string",
          "format": "byte",
          "title": "用户头像"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "title": "创建时间"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "title": "更新时间"
        }
      },
      "title": "用户信息"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}